Migrate DiscountResource to split-class layout
DiscountResource  460 -> 102 lines (+DiscountForm 270, +DiscountTable 90)

Largest Sales resource. Complex form with sections gated on discount
type (BuyXGetY, AmountOff, custom plugin types via
LunarPanelDiscountInterface) plus a Conditions section with
per-currency minimum cart amounts. All granular helpers move to
DiscountForm with the Form suffix dropped.

Cross-resource consumer updated in lockstep:
  ListDiscounts → DiscountForm::getNameComponent / getHandleComponent
                  / getStartsAtComponent / getEndsAtComponent /
                  getDiscountTypeComponent

Closes the Sales family.
