+++ packages/admin/src/Filament/Resources/DiscountResource.php
@@ -2,26 +2,10 @@
 
 namespace Lunar\Admin\Filament\Resources;
 
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Filament\Actions\EditAction;
-use Filament\Forms\Components\DateTimePicker;
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\TextInput;
-use Filament\Forms\Components\Toggle;
 use Filament\Pages\Enums\SubNavigationPosition;
-use Filament\Schemas\Components\Component;
-use Filament\Schemas\Components\Fieldset;
-use Filament\Schemas\Components\Group;
-use Filament\Schemas\Components\Section;
-use Filament\Schemas\Components\Utilities\Get;
-use Filament\Schemas\Components\Utilities\Set;
 use Filament\Schemas\Schema;
 use Filament\Support\Facades\FilamentIcon;
-use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
-use Illuminate\Support\Str;
-use Lunar\Admin\Base\LunarPanelDiscountInterface;
 use Lunar\Admin\Filament\Resources\DiscountResource\Pages\EditDiscount;
 use Lunar\Admin\Filament\Resources\DiscountResource\Pages\ListDiscounts;
 use Lunar\Admin\Filament\Resources\DiscountResource\Pages\ManageDiscountAvailability;
@@ -34,13 +18,10 @@ use Lunar\Admin\Filament\Resources\DiscountResource\RelationManagers\ProductCond
 use Lunar\Admin\Filament\Resources\DiscountResource\RelationManagers\ProductLimitationRelationManager;
 use Lunar\Admin\Filament\Resources\DiscountResource\RelationManagers\ProductRewardRelationManager;
 use Lunar\Admin\Filament\Resources\DiscountResource\RelationManagers\ProductVariantLimitationRelationManager;
+use Lunar\Admin\Filament\Resources\DiscountResource\Schemas\DiscountForm;
+use Lunar\Admin\Filament\Resources\DiscountResource\Tables\DiscountTable;
 use Lunar\Admin\Support\Resources\BaseResource;
-use Lunar\Core\DiscountTypes\AmountOff;
-use Lunar\Core\DiscountTypes\BuyXGetY;
-use Lunar\Core\Facades\Discounts;
 use Lunar\Core\Models\Contracts\Discount as DiscountContract;
-use Lunar\Core\Models\Currency;
-use Lunar\Core\Models\Discount;
 
 class DiscountResource extends BaseResource
 {
@@ -72,359 +53,17 @@ class DiscountResource extends BaseResource
         return __('lunarpanel::global.sections.sales');
     }
 
-    public static function getDefaultForm(Schema $schema): Schema
+    public static function form(Schema $schema): Schema
     {
-        $discountSchemas = Discounts::getTypes()->map(function ($discount) {
-            if (! $discount instanceof LunarPanelDiscountInterface) {
-                return;
-            }
-
-            return Section::make(Str::slug(get_class($discount)))
-                ->heading($discount->getName())
-                ->visible(
-                    fn (Get $get) => $get('type') == get_class($discount)
-                )->schema($discount->lunarPanelSchema());
-        })->filter();
-
-        return $schema->components([
-            Section::make('')->schema(
-                static::getMainFormComponents()
-            ),
-            Section::make('conditions')->schema(
-                static::getConditionsFormComponents()
-            )->heading(
-                __('lunarpanel::discount.form.conditions.heading')
-            ),
-            Section::make('buy_x_get_y')
-                ->heading(
-                    __('lunarpanel::discount.form.buy_x_get_y.heading')
-                )
-                ->visible(
-                    fn (Get $get) => $get('type') == BuyXGetY::class
-                )->schema(
-                    static::getBuyXGetYFormComponents()
-                ),
-            Section::make('amount_off')
-                ->heading(
-                    __('lunarpanel::discount.form.amount_off.heading')
-                )
-                ->visible(
-                    fn (Get $get) => $get('type') == AmountOff::class
-                )->schema(
-                    static::getAmountOffFormComponents()
-                ),
-            ...$discountSchemas,
-        ]);
+        return DiscountForm::configure($schema);
     }
 
-    protected static function getMainFormComponents(): array
+    public static function table(Table $table): Table
     {
-        return [
-            Group::make([
-                static::getNameFormComponent(),
-                static::getHandleFormComponent(),
-            ])->columns(2),
-            Group::make([
-                static::getStartsAtFormComponent(),
-                static::getEndsAtFormComponent(),
-            ])->columns(2),
-            Group::make([
-                static::getPriorityFormComponent(),
-                static::getDiscountTypeFormComponent(),
-            ])->columns(2),
-            static::getStopFormComponent(),
-        ];
+        return DiscountTable::configure($table);
     }
 
-    protected static function getConditionsFormComponents(): array
-    {
-        return [
-            Group::make([
-                static::getCouponFormComponent(),
-                static::getMaxUsesFormComponent(),
-                static::getMaxUsesPerUserFormComponent(),
-            ])->columns(3),
-            Fieldset::make()->schema(
-                static::getMinimumCartAmountsFormComponents()
-            )->label(
-                __('lunarpanel::discount.form.minimum_cart_amount.label')
-            ),
-        ];
-    }
-
-    public static function getNameFormComponent(): Component
-    {
-        return TextInput::make('name')
-            ->label(__('lunarpanel::discount.form.name.label'))
-            ->live(onBlur: true)
-            ->afterStateUpdated(function (string $operation, $state, Set $set) {
-                if ($operation !== 'create') {
-                    return;
-                }
-                $set('handle', Str::slug($state));
-            })
-            ->required()
-            ->maxLength(255)
-            ->autofocus();
-    }
-
-    public static function getHandleFormComponent(): Component
-    {
-        return TextInput::make('handle')
-            ->label(__('lunarpanel::discount.form.handle.label'))
-            ->required()
-            ->unique(ignoreRecord: true)
-            ->live(onBlur: true)
-            ->afterStateUpdated(function (string $operation, $state, Set $set) {
-                if ($operation !== 'create') {
-                    return;
-                }
-
-                $set('handle', Str::snake(Str::lower($state)));
-            })
-            ->maxLength(255)
-            ->autofocus();
-    }
-
-    public static function getStartsAtFormComponent(): Component
-    {
-        return DateTimePicker::make('starts_at')
-            ->label(__('lunarpanel::discount.form.starts_at.label'))
-            ->required()
-            ->before(function (Get $get) {
-                return $get('ends_at');
-            });
-    }
-
-    public static function getEndsAtFormComponent(): Component
-    {
-        return DateTimePicker::make('ends_at')
-            ->label(__('lunarpanel::discount.form.ends_at.label'));
-    }
-
-    protected static function getPriorityFormComponent(): Component
-    {
-        return TextInput::make('priority')
-            ->label(__('lunarpanel::discount.form.priority.label'))
-            ->helperText(
-                __('lunarpanel::discount.form.priority.helper_text')
-            )
-            ->numeric()
-            ->minValue(1)
-            ->maxValue(100)
-            ->default(1);
-    }
-
-    protected static function getStopFormComponent(): Component
-    {
-        return Toggle::make('stop')
-            ->label(
-                __('lunarpanel::discount.form.stop.label')
-            )
-            ->helperText(
-                __('lunarpanel::discount.form.stop.helper_text')
-            );
-    }
-
-    protected static function getCouponFormComponent(): Component
-    {
-        return TextInput::make('coupon')
-            ->label(
-                __('lunarpanel::discount.form.coupon.label')
-            )->helperText(
-                __('lunarpanel::discount.form.coupon.helper_text')
-            )
-            ->unique(ignoreRecord: true);
-    }
-
-    protected static function getMaxUsesFormComponent(): Component
-    {
-        return TextInput::make('max_uses')
-            ->label(
-                __('lunarpanel::discount.form.max_uses.label')
-            )->helperText(
-                __('lunarpanel::discount.form.max_uses.helper_text')
-            );
-    }
-
-    protected static function getMaxUsesPerUserFormComponent(): Component
-    {
-        return TextInput::make('max_uses_per_user')
-            ->label(
-                __('lunarpanel::discount.form.max_uses_per_user.label')
-            )->helperText(
-                __('lunarpanel::discount.form.max_uses_per_user.helper_text')
-            );
-    }
-
-    protected static function getMinimumCartAmountsFormComponents(): array
-    {
-        $currencies = Currency::enabled()->get();
-        $inputs = [];
-
-        foreach ($currencies as $currency) {
-            $inputs[] = TextInput::make('data.min_prices.'.$currency->code)->label(
-                $currency->code
-            )->afterStateHydrated(function (TextInput $component, $state) {
-                $currencyCode = last(explode('.', $component->getStatePath()));
-                $currency = Currency::whereCode($currencyCode)->first();
-
-                if ($currency) {
-                    $component->state($state / $currency->factor);
-                }
-            });
-        }
-
-        return $inputs;
-    }
-
-    public static function getDiscountTypeFormComponent(): Component
-    {
-        return Select::make('type')->options(
-            Discounts::getTypes()->mapWithKeys(
-                fn ($type) => [get_class($type) => $type->getName()]
-            )
-        )->required()->live();
-    }
-
-    protected static function getAmountOffFormComponents(): array
-    {
-        $currencies = Currency::get();
-
-        $currencyInputs = [];
-
-        foreach ($currencies as $currency) {
-            $currencyInputs[] = TextInput::make(
-                'data.fixed_values.'.$currency->code
-            )->label($currency->name)->afterStateHydrated(function (TextInput $component, $state) use ($currencies) {
-                $currencyCode = last(explode('.', $component->getStatePath()));
-                $currency = $currencies->first(
-                    fn ($currency) => $currency->code == $currencyCode
-                );
-
-                if ($currency) {
-                    $component->state($state / $currency->factor);
-                }
-            });
-        }
-
-        return [
-            Toggle::make('data.fixed_value')
-                ->label(__('lunarpanel::discount.form.fixed_value.label'))
-                ->live(),
-            TextInput::make('data.percentage')
-                ->label(__('lunarpanel::discount.form.percentage.label'))
-                ->visible(
-                    fn (Get $get) => ! $get('data.fixed_value')
-                )->numeric(),
-            Group::make(
-                $currencyInputs
-            )->visible(
-                fn (Get $get) => (bool) $get('data.fixed_value')
-            )->columns(3),
-        ];
-    }
-
-    public static function getBuyXGetYFormComponents(): array
-    {
-        return [
-            TextInput::make('data.min_qty')
-                ->label(
-                    __('lunarpanel::discount.form.min_qty.label')
-                )->helperText(
-                    __('lunarpanel::discount.form.min_qty.helper_text')
-                )->numeric(),
-            Group::make([
-                TextInput::make('data.reward_qty')
-                    ->label(
-                        __('lunarpanel::discount.form.reward_qty.label')
-                    )->helperText(
-                        __('lunarpanel::discount.form.reward_qty.helper_text')
-                    )->numeric(),
-                TextInput::make('data.max_reward_qty')
-                    ->label(
-                        __('lunarpanel::discount.form.max_reward_qty.label')
-                    )->helperText(
-                        __('lunarpanel::discount.form.max_reward_qty.helper_text')
-                    )->numeric(),
-            ])->columns(2),
-            Toggle::make('data.automatically_add_rewards')
-                ->label(
-                    __('lunarpanel::discount.form.automatic_rewards.label')
-                )->helperText(
-                    __('lunarpanel::discount.form.automatic_rewards.helper_text')
-                ),
-        ];
-    }
-
-    public static function getDefaultTable(Table $table): Table
-    {
-        return $table
-            ->columns(static::getTableColumns())
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ])->searchable();
-    }
-
-    protected static function getTableColumns(): array
-    {
-        return [
-            TextColumn::make('status')
-                ->formatStateUsing(function ($state) {
-                    return __("lunarpanel::discount.table.status.{$state}.label");
-                })
-                ->label(__('lunarpanel::discount.table.status.label'))
-                ->badge()
-                ->color(fn (string $state): string => match ($state) {
-                    Discount::ACTIVE => 'success',
-                    Discount::EXPIRED => 'danger',
-                    Discount::PENDING => 'gray',
-                    Discount::SCHEDULED => 'info',
-                })
-                ->toggleable(),
-            TextColumn::make('name')
-                ->label(__('lunarpanel::discount.table.name.label'))
-                ->searchable()
-                ->sortable()
-                ->toggleable(),
-            TextColumn::make('type')
-                ->formatStateUsing(function ($state) {
-                    return (new $state)->getName();
-                })
-                ->label(__('lunarpanel::discount.table.type.label'))
-                ->toggleable(),
-            TextColumn::make('starts_at')
-                ->label(__('lunarpanel::discount.table.starts_at.label'))
-                ->date()
-                ->sortable()
-                ->toggleable(),
-            TextColumn::make('ends_at')
-                ->label(__('lunarpanel::discount.table.ends_at.label'))
-                ->date()
-                ->sortable()
-                ->toggleable(),
-            TextColumn::make('coupon')
-                ->label(__('lunarpanel::discount.table.coupon.label'))
-                ->searchable()
-                ->sortable()
-                ->toggleable(isToggledHiddenByDefault: true),
-            TextColumn::make('created_at')
-                ->label(__('lunarpanel::discount.table.created_at.label'))
-                ->date()
-                ->sortable()
-                ->toggleable(isToggledHiddenByDefault: true),
-        ];
-    }
-
-    public static function getDefaultSubNavigation(): array
+    protected static function getDefaultSubNavigation(): array
     {
         return [
             EditDiscount::class,
@@ -447,7 +86,7 @@ class DiscountResource extends BaseResource
         ];
     }
 
-    public static function getDefaultPages(): array
+    protected static function getDefaultPages(): array
     {
         return [
             'index' => ListDiscounts::route('/'),
