+++ packages/admin/src/Filament/Resources/DiscountResource/Tables/DiscountTable.php
@@ -1,0 +1,85 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\DiscountResource\Tables;
+
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+use Lunar\Core\Models\Discount;
+
+class DiscountTable
+{
+    use CallsHooks;
+
+    public static function configure(Table $table): Table
+    {
+        return self::callStaticLunarHook(
+            'configureTable',
+            $table
+                ->columns(static::getColumns())
+                ->filters([])
+                ->recordActions([
+                    EditAction::make(),
+                ])
+                ->toolbarActions([
+                    BulkActionGroup::make([
+                        DeleteBulkAction::make(),
+                    ]),
+                ])
+                ->searchable(),
+        );
+    }
+
+    public static function getColumns(): array
+    {
+        return [
+            TextColumn::make('status')
+                ->formatStateUsing(function ($state) {
+                    return __("lunarpanel::discount.table.status.{$state}.label");
+                })
+                ->label(__('lunarpanel::discount.table.status.label'))
+                ->badge()
+                ->color(fn (string $state): string => match ($state) {
+                    Discount::ACTIVE => 'success',
+                    Discount::EXPIRED => 'danger',
+                    Discount::PENDING => 'gray',
+                    Discount::SCHEDULED => 'info',
+                })
+                ->toggleable(),
+            TextColumn::make('name')
+                ->label(__('lunarpanel::discount.table.name.label'))
+                ->searchable()
+                ->sortable()
+                ->toggleable(),
+            TextColumn::make('type')
+                ->formatStateUsing(function ($state) {
+                    return (new $state)->getName();
+                })
+                ->label(__('lunarpanel::discount.table.type.label'))
+                ->toggleable(),
+            TextColumn::make('starts_at')
+                ->label(__('lunarpanel::discount.table.starts_at.label'))
+                ->date()
+                ->sortable()
+                ->toggleable(),
+            TextColumn::make('ends_at')
+                ->label(__('lunarpanel::discount.table.ends_at.label'))
+                ->date()
+                ->sortable()
+                ->toggleable(),
+            TextColumn::make('coupon')
+                ->label(__('lunarpanel::discount.table.coupon.label'))
+                ->searchable()
+                ->sortable()
+                ->toggleable(isToggledHiddenByDefault: true),
+            TextColumn::make('created_at')
+                ->label(__('lunarpanel::discount.table.created_at.label'))
+                ->date()
+                ->sortable()
+                ->toggleable(isToggledHiddenByDefault: true),
+        ];
+    }
+}
